feat: add unit field to products and decimal quantity for sale items
- Add Unit property to Product entity (default: 'unidad')
- Change SaleItem.Quantity from int to decimal (supports kg, lb, g, etc.)
- Update all DTOs: ProductDto, CreateProductDto, UpdateProductDto, SyncProductDto, SyncSaleItemDto, TopProductDto
- Update AppDbContext precision config for Quantity (18,3) and Unit column
- Update SyncController and ProductsController to map Unit field
- Add EF Core migration: AddUnitAndDecimalQuantity

+++ Src/POSI.Api/Controllers/SalesController.cs
@@ -43,7 +43,7 @@ public class SalesController : ControllerBase
             if (item.ProductId is null) continue;
             var product = await _db.Products.FindAsync(item.ProductId.Value);
             if (product is not null)
-                product.Stock += item.Quantity;
+                product.Stock += (int)item.Quantity;
         }
 
         await _db.SaveChangesAsync();
